refactor(runtime): extract ai sdk turn execution
Generated-by: Codex
